requirements.txt
tensorflow>=2.10
numpy
matplotlib
scikit-learn
pillow
streamlit
tqdm
